+++ packages/layout/src/array.ts
@@ -1,4 +1,4 @@
-import { SOLIDITY_TYPES } from "types";
+import { SOLIDITY_TYPES } from "@forge-jig/types";
 import { StorageInfoStruct } from "./types";
 
 export interface ArrayPointer {