+++ packages/layout/src/mapping.ts
@@ -1,4 +1,4 @@
-import { SOLIDITY_TYPES } from "types";
+import { SOLIDITY_TYPES } from "@forge-jig/types";
 import { ArrayPointer } from "./array";
 import { isMappingPointer } from "./predicate";
 import {