+++ packages/codegen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # codegen
 
+## 0.1.3
+
+### Patch Changes
+
+- move packages under scoped name
+
 ## 0.1.2
 
 ### Patch Changes