+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # config
 
+## 1.0.3
+
+### Patch Changes
+
+- move packages under scoped name
+
 ## 1.0.2
 
 ### Patch Changes